backend/requirements.txt
fastapi[standard]
sqlmodel
python-dotenv
auth0-fastapi
